requirements.txt
beautifulsoup4==4.10.0
numpy==1.22.2
pandas==1.4.1
parse==1.19.0
plotly-express==0.4.0
pycparser==2.21
pyparsing==3.0.7
requests==2.27.1
soupsieve==2.3.1
streamlit==1.6.0
XLML==0.1.2
lxml==4.8.0
